fix: duplication in activity
